Separate unit tests into areas of concern
Replaces the single monolithic unit test suite into five separate files.

* A client test suite
* A perform-request test suite
* A request-endpoint test suite
* A helper test suite
* An endpoint-specific test suite
